requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
pyyaml>=6.0
python-dotenv>=1.0
jinja2>=3.1

# ─── HTTP / Feeds ─────────────────────────────────────────────────────────────
requests>=2.31
aiohttp>=3.9
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=5.0             # faster feedparser backend

# ─── Sources ──────────────────────────────────────────────────────────────────
praw>=7.7             # Reddit
arxiv>=2.1            # arXiv (no API key needed)
google-api-python-client>=2.100  # YouTube

# ─── Scheduling ───────────────────────────────────────────────────────────────
APScheduler>=3.10

# ─── LLM (optional — only needed when llm.enabled=true) ──────────────────────
anthropic>=0.40       # Claude
# openai>=1.0         # uncomment if using OpenAI instead

# ─── Semantic matching (optional — improves relevance scoring) ─────────────────
# Uncomment to enable semantic/embedding-based relevance matching:
# sentence-transformers>=3.0
# torch>=2.0           # required by sentence-transformers

# ─── Google Scholar (optional — rate-limited) ─────────────────────────────────
# scholarly>=1.7       # uncomment if google_scholar source is enabled

# ─── Airflow (optional — only if using the Airflow DAG) ──────────────────────
# apache-airflow>=2.8  # uncomment if deploying via Airflow
